Moved all opennlp stuff to opennlp package.
